refactor: isolate runner connection lifecycle
