+++ src/core/error/mod.rs
@@ -35,7 +35,6 @@ pub enum ErrorCode {
     RigPipelineFailed,
     RigServiceFailed,
     RigResourceConflict,
-    RunnerNotFound,
     StackNotFound,
     StackApplyConflict,
 
@@ -85,7 +84,6 @@ impl ErrorCode {
             ErrorCode::RigPipelineFailed => "rig.pipeline_failed",
             ErrorCode::RigServiceFailed => "rig.service_failed",
             ErrorCode::RigResourceConflict => "rig.resource_conflict",
-            ErrorCode::RunnerNotFound => "runner.not_found",
             ErrorCode::StackNotFound => "stack.not_found",
             ErrorCode::StackApplyConflict => "stack.apply_conflict",
